Moved mail helper classes into separate files
