+++ sendgrid/__init__.py
@@ -1,4 +1,4 @@
 from .version import __version__  # noqa
 # v3 API
 from .sendgrid import SendGridAPIClient  # noqa
-from .helpers.mail.mail import Email  # noqa
+from .helpers.mail import Email  # noqa
