+++ sendgrid/helpers/mail/category.py
@@ -1,0 +1,17 @@
+class Category(object):
+
+    def __init__(self, name=None):
+        self._name = None
+        if name is not None:
+            self._name = name
+
+    @property
+    def name(self):
+        return self._name
+
+    @name.setter
+    def name(self, value):
+        self._name = value
+
+    def get(self):
+        return self.name
