+++ sendgrid/helpers/mail/ganalytics.py
@@ -1,0 +1,92 @@
+class Ganalytics(object):
+
+    def __init__(self,
+                 enable=None,
+                 utm_source=None,
+                 utm_medium=None,
+                 utm_term=None,
+                 utm_content=None,
+                 utm_campaign=None):
+        self._enable = None
+        self._utm_source = None
+        self._utm_medium = None
+        self._utm_term = None
+        self._utm_content = None
+        self._utm_campaign = None
+
+        if enable is not None:
+            self.enable = enable
+        if utm_source is not None:
+            self.utm_source = utm_source
+        if utm_medium is not None:
+            self.utm_medium = utm_medium
+        if utm_term is not None:
+            self.utm_term = utm_term
+        if utm_content is not None:
+            self.utm_content = utm_content
+        if utm_campaign is not None:
+            self.utm_campaign = utm_campaign
+
+    @property
+    def enable(self):
+        return self._enable
+
+    @enable.setter
+    def enable(self, value):
+        self._enable = value
+
+    @property
+    def utm_source(self):
+        return self._utm_source
+
+    @utm_source.setter
+    def utm_source(self, value):
+        self._utm_source = value
+
+    @property
+    def utm_medium(self):
+        return self._utm_medium
+
+    @utm_medium.setter
+    def utm_medium(self, value):
+        self._utm_medium = value
+
+    @property
+    def utm_term(self):
+        return self._utm_term
+
+    @utm_term.setter
+    def utm_term(self, value):
+        self._utm_term = value
+
+    @property
+    def utm_content(self):
+        return self._utm_content
+
+    @utm_content.setter
+    def utm_content(self, value):
+        self._utm_content = value
+
+    @property
+    def utm_campaign(self):
+        return self._utm_campaign
+
+    @utm_campaign.setter
+    def utm_campaign(self, value):
+        self._utm_campaign = value
+
+    def get(self):
+        ganalytics = {}
+        if self.enable is not None:
+            ganalytics["enable"] = self.enable
+        if self.utm_source is not None:
+            ganalytics["utm_source"] = self.utm_source
+        if self.utm_medium is not None:
+            ganalytics["utm_medium"] = self.utm_medium
+        if self.utm_term is not None:
+            ganalytics["utm_term"] = self.utm_term
+        if self.utm_content is not None:
+            ganalytics["utm_content"] = self.utm_content
+        if self.utm_campaign is not None:
+            ganalytics["utm_campaign"] = self.utm_campaign
+        return ganalytics
