+++ sendgrid/helpers/mail/sandbox_mode.py
@@ -1,0 +1,21 @@
+class SandBoxMode(object):
+
+    def __init__(self, enable=None):
+        self._enable = None
+
+        if enable is not None:
+            self.enable = enable
+
+    @property
+    def enable(self):
+        return self._enable
+
+    @enable.setter
+    def enable(self, value):
+        self._enable = value
+
+    def get(self):
+        sandbox_mode = {}
+        if self.enable is not None:
+            sandbox_mode["enable"] = self.enable
+        return sandbox_mode
