+++ sendgrid/helpers/mail/subscription_tracking.py
@@ -1,0 +1,63 @@
+class SubscriptionTracking(object):
+
+    def __init__(self, enable=None, text=None, html=None, substitution_tag=None):
+        self._enable = None
+        self._text = None
+        self._html = None
+        self._substitution_tag = None
+
+        if enable is not None:
+            self.enable = enable
+        if text is not None:
+            self.text = text
+        if html is not None:
+            self.html = html
+        if substitution_tag is not None:
+            self.substitution_tag = substitution_tag
+
+    @property
+    def enable(self):
+        return self._enable
+
+    @enable.setter
+    def enable(self, value):
+        self._enable = value
+
+    @property
+    def text(self):
+        return self._text
+
+    @text.setter
+    def text(self, value):
+        self._text = value
+
+    @property
+    def html(self):
+        return self._html
+
+    @html.setter
+    def html(self, value):
+        self._html = value
+
+    @property
+    def substitution_tag(self):
+        return self._substitution_tag
+
+    @substitution_tag.setter
+    def substitution_tag(self, value):
+        self._substitution_tag = value
+
+    def get(self):
+        subscription_tracking = {}
+        if self.enable is not None:
+            subscription_tracking["enable"] = self.enable
+
+        if self.text is not None:
+            subscription_tracking["text"] = self.text
+
+        if self.html is not None:
+            subscription_tracking["html"] = self.html
+
+        if self.substitution_tag is not None:
+            subscription_tracking["substitution_tag"] = self.substitution_tag
+        return subscription_tracking
